+++ packages/apollo-utilities/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ### vNext
+- BREAKING: Add `typename` as a required second argument in `toIdValue` helper [PR#3159](https://github.com/apollographql/apollo-client/pull/3159)
 - Map coverage to original source
 - Added ability to include or ignore fragments in `getDirectivesFromDocument` [PR#3010](https://github.com/apollographql/apollo-client/pull/3010)